+++ lib/prefixed_ids.rb
@@ -38,11 +38,10 @@ module PrefixedIds
     end
 
     class_methods do
-      def has_prefix_id(prefix, override_find: true, override_param: true, override_exists: true, fallback: true, **options)
+      def has_prefix_id(prefix, override_find: true, override_param: true, fallback: true, **options)
         include Attribute
         include Finder if override_find
         include ToParam if override_param
-        include Exists if override_exists
         self._prefix_id = PrefixId.new(self, prefix, **options)
         self._prefix_id_fallback = fallback
 
@@ -122,20 +121,6 @@ module PrefixedIds
 
     def to_param
       _prefix_id.encode(id)
-    end
-  end
-
-  module Exists
-    extend ActiveSupport::Concern
-
-    class_methods do
-      def exists?(id)
-        if _prefix_id.present? && id.is_a?(String)
-          super(_prefix_id.decode(id))
-        else
-          super
-        end
-      end
     end
   end
 end